Replaced invalid ErrorAction parameter value
On a fresh machine where Neo4J has never been installed (and has no corresponding registry entries) the installation fails because `Ignore` is passed to the `ErrorAction` parameter when the install script looks for the existing Neo4j home path (and later for the installed service).  The error happens because `Ignore` is not actually defined in the corresponding enum `System.Management.Automation.ActionPreference`.  Valid alternatives with the similar intention are: `Continue` and `SilentlyContinue`.  Here, two occurrences of `Ignore` were replaced with `Continue`.